Move CatalogItem to catalog-rest-api
