+++ rest-api/src/main/java/org/brooklyncentral/catalog/dto/CatalogItem.java
@@ -1,4 +1,4 @@
-package org.brooklyncentral.catalog.model;
+package org.brooklyncentral.catalog.dto;
 
 import java.util.Map;
